Add files via upload
We have name the project as a Day and Night Cycle because we can change the mode from Day to Night and vice versa .In this mini project, we have developed some objects like a house , sun , clouds , boats , trees and birds. Couple of clouds will be passing in the sky and the set of birds will be flying in the sky . It has got two modes in it, Day Mode and the Night Mode, the above explained things are attained in the day mode and in night mode moon , clouds will be passing in the Night sky. The code implemented makes use of various OpenGL functions for translation, keyboard callback function, built-in functions for textures and many more. The concepts of computer graphics and OpenGL stand a backbone to achieve the aforementioned idea. Primitive drawing, event driven interactions and basic animation have been the important concepts brought out by this application.